Fix: PlayerStateController, PlayerAnimator, delete JUMP and CLIMB
