+++ 1984/Assets/Scenes/Game_Platformer.unity
@@ -1472,7 +1472,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 682191246}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -65.86197, y: 1.8768201, z: -8.9}
+  m_LocalPosition: {x: -38.17005, y: 1.8631799, z: -8.9}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -2196,6 +2196,10 @@ PrefabInstance:
       propertyPath: m_Materials.Array.data[0]
       value: 
       objectReference: {fileID: 1354637102}
+    - target: {fileID: 2343838597052046403, guid: f1b8e01b1aac184438e8ab376af23aeb, type: 3}
+      propertyPath: m_Enabled
+      value: 1
+      objectReference: {fileID: 0}
     - target: {fileID: 3359077727555832881, guid: f1b8e01b1aac184438e8ab376af23aeb, type: 3}
       propertyPath: m_Materials.Array.data[0]
       value:
