+++ 1984/Assets/Scripts/Runtime/Map/Telescreen/ITelescreenState.cs
