+++ 1984/Assets/Scripts/Runtime/Player/Control/IdleState.cs
@@ -9,12 +9,8 @@ public class IdleState : IState
     private PlayerStateController stateController;
     private PlayerMoveTracker moveTracker;
     private Rigidbody2D rigidbody;
-    private float keyVertical;
     private float keyHorizontal;
-    private bool keySpace;
-    private GameObject ladder;
     public Vector3 playerPos;
-    public Vector3 ladderPos;
 
     public IdleState(PlayerStateController playerStateController, PlayerMoveTracker playerMoveTracker, Rigidbody2D rigidbody2D)
     {
@@ -35,15 +31,7 @@ public class IdleState : IState
     public void Execute()
     {
         //Debug.Log("Idle Execute()");
-        keyVertical = Input.GetAxisRaw("Vertical");
         keyHorizontal = Input.GetAxisRaw("Horizontal");
-        keySpace = Input.GetKey(KeyCode.Space);
-
-        //Idle -> Jump
-        if (keySpace && moveTracker.isGrounded)
-        {
-            stateController.ChangeState(PLAYER_STATE.JUMP);
-        }
     }
     public void FixedExecute()
     {
@@ -56,25 +44,10 @@ public class IdleState : IState
             stateController.ChangeState(PLAYER_STATE.RUN);
         }
 
-        if (keyVertical != 0 && moveTracker.isNearLadder)
-        {
-            ladder = moveTracker.ladderObj;
-            ladderPos = ladder.transform.position;
-
-            //Idle -> Climb
-            //down
-            if (keyVertical < 0 && playerPos.y > ladderPos.y)
-                stateController.ChangeState(PLAYER_STATE.CLIMB);
-             
-            //up
-            if (keyVertical > 0 && playerPos.y < ladderPos.y)
-                stateController.ChangeState(PLAYER_STATE.CLIMB);
-        }
     }
 
     public void Exit()
     {
-        //Debug.Log("Idle Exit()");
-        
+        //Debug.Log("Idle Exit()");       
     }
 }
