+++ 1984/Assets/Scripts/Runtime/Player/Control/RunState.cs
@@ -8,14 +8,10 @@ public class RunState : IState
     private PlayerStateController stateController;
     private PlayerMoveTracker moveTracker;
     private Rigidbody2D rigidbody;
-    private readonly float moveSpeed = 20f;
     private float keyHorizontal;
-    private float keyVertical;
     private Vector3 playerPos;
     private Vector3 playerScale;
-
-    private GameObject ladder;
-    public Vector3 ladderPos;
+    private float speed;
 
     public RunState(PlayerStateController playerStateController, PlayerMoveTracker playerMoveTracker, Rigidbody2D rigidbody2D)
     {
@@ -27,6 +23,8 @@ public class RunState : IState
     public void Enter()
     {
         //Debug.Log("Run Enter()");
+        speed = moveTracker.runSpeed;
+
         playerPos = stateController.transform.position;
         stateController.transform.localScale = moveTracker.isRight? new Vector3(1f, 1f, 1f): new Vector3(-1f, 1f, 1f);
         playerScale = new Vector3(1f, 1f, 1f);
@@ -35,14 +33,7 @@ public class RunState : IState
     public void Execute()
     {
         //Debug.Log("Run Execute()");
-        keyVertical = Input.GetAxisRaw("Vertical");
         keyHorizontal = Input.GetAxisRaw("Horizontal");
-
-        //Run -> Jump
-        if (Input.GetKey(KeyCode.Space) && moveTracker.isGrounded)
-        {
-            stateController.ChangeState(PLAYER_STATE.JUMP);
-        }
     }
 
     public void FixedExecute()
@@ -50,7 +41,7 @@ public class RunState : IState
         if (keyHorizontal != 0)
         {
             //Run           
-            rigidbody.velocity = new Vector2(keyHorizontal * moveSpeed, rigidbody.velocity.y);
+            rigidbody.velocity = new Vector2(keyHorizontal * speed, rigidbody.velocity.y);
         }
 
         else
@@ -60,20 +51,6 @@ public class RunState : IState
             stateController.ChangeState(PLAYER_STATE.IDLE);
         }
 
-        if (keyVertical != 0 && moveTracker.isNearLadder)
-        {
-            ladder = moveTracker.ladderObj;
-            ladderPos = ladder.transform.position;
-
-            //Climb down
-            if (keyVertical < 0 && playerPos.y > ladderPos.y)
-                stateController.ChangeState(PLAYER_STATE.CLIMB);
-
-            //Climb up
-            if (keyVertical > 0 && playerPos.y < ladderPos.y)
-                stateController.ChangeState(PLAYER_STATE.CLIMB);
-
-        }
     }
 
     public void Exit()
