--- 1984/Assets/Scripts/Runtime/UI/Platformer/ClimbState.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b85f3f582edcf964f8fe047a82908328
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
