--- 1984/Assets/Scripts/Runtime/UI/Platformer/JumpState.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0cde203e0af1a724f90b4909aaa32873
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
