--- 1984/Assets/Scripts/Runtime/UI/Platformer/PlayerMoveTracker.cs
@@ -1,54 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
- 
-public class PlayerMoveTracker : MonoBehaviour
-{
-    public bool isNearLadder;
-    public bool isGrounded;
-    public bool isRight;
-    public bool isTraces;
-    [HideInInspector]
-    public GameObject ladderObj;
-    private CapsuleCollider2D coll;
-    private LayerMask groundLayer;
-     
-
-    private void Start()
-    { 
-        coll = GetComponent<CapsuleCollider2D>(); 
-        groundLayer = LayerMask.GetMask("Ground");
-        isRight = true;
-
-    }
-
-    private void Update()
-    {
-        CheckGrounded();    
-    }
-
-    private void OnTriggerEnter2D(Collider2D other)
-    {
-        if (other.gameObject.CompareTag("Ladder"))
-        {
-            isNearLadder = true;
-            ladderObj = other.gameObject;
-        }
-    }
-
-    private void OnTriggerExit2D(Collider2D other)
-    {
-        if (other.gameObject.CompareTag("Ladder"))
-        {
-            isNearLadder = false;
-            ladderObj = null;
-        }
-    }
-
-    private void CheckGrounded()
-    {
-        Bounds bounds = coll.bounds;
-        Vector3 footPosition = new Vector2(bounds.center.x, bounds.min.y);
-        isGrounded = Physics2D.OverlapCircle(footPosition, 0.1f, groundLayer) ? true : false;
-    }
-}
